refactor: remove typescript support
Reasoning behind this decision is that this CLI tool is built for
use case that we encountered at work. We have legacy application
built using jquery. Our goal is to slowly migrate it to Vue and
we think that typescript is not the tool that will help us at
this. Some JavaScript vs TypeScript:

- We can add 'types' to a codebase with JSDoc comments
- Some part of our applications are already migrated to Vue.js 2,
  so its not worth upgrading that to TypeScript
- If this was a new project, starting from zero, we would probably
  add TypeScript support.